+++ src/core/utils.rs
@@ -1,5 +1,5 @@
 /// Estimates text width using character categorization.
-pub(crate) fn estimate_text_width(text: &str, font_size: f32) -> f32 {
+pub(crate) fn estimate_text_width(text: &str, font_size: f64) -> f64 {
     let mut narrow_chars = 0;
     let mut uppercase_chars = 0;
     let mut other_chars = 0;
@@ -14,5 +14,5 @@ pub(crate) fn estimate_text_width(text: &str, font_size: f32) -> f32 {
         }
     }
 
-    (narrow_chars as f32 * 0.3 + uppercase_chars as f32 * 0.65 + other_chars as f32 * 0.55) * font_size
+    (narrow_chars as f64 * 0.3 + uppercase_chars as f64 * 0.65 + other_chars as f64 * 0.55) * font_size
 }
