+++ src/mark/rect.rs
@@ -9,9 +9,9 @@ use crate::visual::color::SingleColor;
 #[derive(Clone)]
 pub struct MarkRect {
     pub(crate) color: SingleColor,
-    pub(crate) opacity: f32,
+    pub(crate) opacity: f64,
     pub(crate) stroke: SingleColor,
-    pub(crate) stroke_width: f32,
+    pub(crate) stroke_width: f64,
 }
 
 impl MarkRect {
@@ -36,7 +36,7 @@ impl MarkRect {
     /// Sets the opacity of the rectangle mark.
     /// 
     /// Value should be between 0.0 (transparent) and 1.0 (opaque).
-    pub fn with_opacity(mut self, opacity: f32) -> Self {
+    pub fn with_opacity(mut self, opacity: f64) -> Self {
         self.opacity = opacity.clamp(0.0, 1.0);
         self
     }
@@ -48,7 +48,7 @@ impl MarkRect {
     }
 
     /// Sets the thickness of the rectangle's outline.
-    pub fn with_stroke_width(mut self, width: f32) -> Self {
+    pub fn with_stroke_width(mut self, width: f64) -> Self {
         self.stroke_width = width;
         self
     }
